chore: refactor to use tool layout
