--- tools/json-formatter/src/components/AdBanner.vue
@@ -1,11 +1,0 @@
-<template>
-	<div class="page" >
-		<b-jumbotron bg-variant="info" header="ADS GO HERE"/>
-	</div>
-</template>
-
-<script>
-export default {
-	name: 'AdBanner',
-}
-</script>
