+++ tools/json-formatter/src/components/JsonFormatter.vue
